Refactor the ExtensionRow to make it possible to reuse more elements for the AppHomeRow
