+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/ExtensionRow/ExtensionRow.module.scss
@@ -1,13 +1,3 @@
-
-.DevToolRow {
-  border-bottom: 1px solid var(--dc-border);
-
-  td {
-    padding: 0.8rem 1.5rem;
-    vertical-align: top;
-  }
-}
-
 .Title {
   font-weight: var(--dc-font-weight-subdued);
 }
