+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/ExtensionRow/components/PreviewLinks/translations/en.json
@@ -1,8 +1,3 @@
 {
-  "copy": {
-    "success": "URL copied",
-    "error": "Couldn't copy URL"
-  },
-  "linkLabel": "Preview {title}",
-  "iconLabel": "Copy preview link for {title} to clipboard"
+  "previewLinksTitle": "Multiple locations"
 }
