+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/ExtensionRow/translations/en.json
@@ -1,4 +1,7 @@
 {
+  "viewMobile": "View mobile",
+  "local": "Local",
+  "live": "Live",
   "statuses": {
     "success": "Connected",
     "disconnected": "Disconnected",
