+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/NotApplicable/NotApplicable.module.scss
@@ -1,0 +1,3 @@
+.NotApplicable {
+  font-family: var(--dc-font-compact);
+}
