+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/NotApplicable/NotApplicable.tsx
@@ -1,0 +1,6 @@
+import * as styles from './NotApplicable.module.scss'
+import React from 'react'
+
+export function NotApplicable() {
+  return <span className={styles.NotApplicable}>--</span>
+}
