+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/PreviewLink/index.ts
@@ -1,0 +1,1 @@
+export * from './PreviewLink'
