+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/PreviewLink/translations/en.json
@@ -1,0 +1,8 @@
+{
+  "copy": {
+    "success": "URL copied",
+    "error": "Couldn't copy URL"
+  },
+  "linkLabel": "Preview {title}",
+  "iconLabel": "Copy preview link for {title} to clipboard"
+}
