+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/Row/Row.module.scss
@@ -1,0 +1,8 @@
+.Row {
+  border-bottom: 1px solid var(--dc-border);
+
+  td {
+    padding: 0.8rem 1.5rem;
+    vertical-align: top;
+  }
+}
