+++ packages/ui-extensions-dev-console/src/sections/Extensions/components/index.ts
@@ -1,0 +1,5 @@
+export * from './ExtensionRow'
+export * from './NotApplicable'
+export * from './PreviewLink'
+export * from './QRCodeModal'
+export * from './Row'
